requirements.txt
deta==1.1.0
keras==2.9.0
matplotlib==3.5.2
numpy==1.22.3
opencv_python==4.6.0.66
pandas==1.4.3
Pillow==9.3.0
PyAutoGUI==0.9.53
pydicom==2.3.0
scikit_learn==1.1.3
scipy==1.7.3
streamlit==1.14.0
tensorflow==2.9.1
